8-models/requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
